+++ crates/parse_selectors/tests/js/location/output.js
@@ -33,3 +33,6 @@ window.location =foo;
 window.location= foo;
 window.location=foo;
 window.location = foo.bar;
+
+window.location.hash = "#a";
+window.location.hash = "#a";