+++ crates/parse_selectors/tests/js/location/source.js
@@ -33,3 +33,6 @@ window.location =foo;
 window.location= foo;
 window.location=foo;
 window.location = foo.bar;
+
+window.location.hash = "#SELECTOR-%31";
+window.location.hash = "#%53ELECTOR-1";